Improve UI * implement search * hookup dropdowns * add search field

+++ console/model.go
@@ -5,13 +5,13 @@ import (
 )
 
 type LogRun struct {
-	ID        int64     `db:"id" json:"id"`
+	ID        int       `db:"id" json:"id"`
 	CreatedAt time.Time `db:"created_at" json:"createdAt"`
 }
 
 type LogEvent struct {
-	ID        int64     `db:"id" json:"id"`
-	RunID     int64     `db:"run_id" json:"runId"`
+	ID        int       `db:"id" json:"id"`
+	RunID     int       `db:"run_id" json:"runId"`
 	EventType string    `db:"event_type" json:"eventType"`
 	EventData string    `db:"event_data" json:"eventData"`
 	CreatedAt time.Time `db:"created_at" json:"createdAt"`
